refactor(scan): split runtime universe selections
